refactor(backend): move toLearningStatsModel out of mapper.go into stats resolver (#869)
